--- src/components/Tooltip/TooltipHeader.tsx
@@ -1,14 +1,0 @@
-export const TooltipHeader = () => {
-  return (
-    <div
-      style={{
-        fontSize: '1.6rem',
-        fontWeight: '600',
-        lineHeight: '2.4rem',
-        color: '#333333',
-      }}
-    >
-      ToolTip Header
-    </div>
-  );
-};
